# swagger.yaml
openapi: 3.0.0
info:
  title: UberEATS API
  description: API documentation for UberEATS prototype
  version: 1.0.0
servers:
  - url: http://localhost:3001/api
    description: Local development server
...
paths:
...
  /customer/profile:
    get:
      tags:
        - Customer
      summary: Get customer profile
      description: Get logged in customer's profile details
      responses:
        '200':
          description: Customer profile data
        '401':
          description: Unauthorized
        '404':
          description: Profile not found
        '500':
          description: Server error
    put:
      tags:
        - Customer
      summary: Update customer profile
      description: Update logged in customer's profile information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone:
                  type: string
                  example: '1234567890'
                address:
                  type: string
                  example: '123 Main St'
                country:
                  type: string
                  example: 'USA'
                state:
                  type: string
                  example: 'CA'
                city:
                  type: string
                  example: 'San Jose'
      responses:
        '200':
          description: Profile updated successfully
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '404':
          description: Profile not found
        '500':
          description: Server error
...
  /restaurants:
    get:
      tags:
        - Restaurant
      summary: Get all restaurants
      description: Get a list of all restaurants
      responses:
        '200':
          description: List of restaurants
        '500':
          description: Server error
  /restaurants/{id}:
    get:
      tags:
        - Restaurant
      summary: Get restaurant by ID
      description: Get details of a specific restaurant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 2
      responses:
        '200':
          description: Restaurant details
        '404':
          description: Restaurant not found
        '500':
          description: Server error
...
  /restaurant/profile:
    get:
      tags:
        - Restaurant
      summary: Get restaurant profile
      description: Get logged in restaurant's profile
      responses:
        '200':
          description: Restaurant profile data
        '401':
          description: Unauthorized
        '404':
          description: Profile not found
        '500':
          description: Server error
    put:
      tags:
        - Restaurant
      summary: Update restaurant profile
      description: Update logged in restaurant's profile information
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                  example: 'Authentic Italian cuisine'
                location:
                  type: string
                  example: '123 Main St, San Jose, CA'
                cuisine:
                  type: string
                  example: 'Italian'
                openTime:
                  type: string
                  format: time
                  example: '11:00'
                closeTime:
                  type: string
                  format: time
                  example: '22:00'
                deliveryFee:
                  type: number
                  example: 2.99
                minOrder:
                  type: number
                  example: 15
      responses:
        '200':
          description: Profile updated successfully
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '404':
          description: Profile not found
        '500':
          description: Server error
  /restaurant/profile/image:
    post:
      tags:
        - Restaurant
      summary: Upload restaurant image
      description: Upload an image for the restaurant
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
      responses:
        '200':
          description: Restaurant image updated successfully
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '404':
          description: Profile not found
        '500':
          description: Server error
...